refactor(transport): extract FileRegistry tests to sibling module
file_registry.rs (759 lines) drops to 463 lines by moving the 298-line
#[cfg(test)] mod tests block into a sibling file_registry_tests.rs
mounted via #[cfg(test)] #[path = "file_registry_tests.rs"] mod tests;

Struct fields are private and method-impl blocks must co-locate with
the struct to access them, so the impl itself stays in the facade.
No behaviour change; 130 + 7 + 3 transport tests continue to pass.
